requirements.txt
aiohttp>=3.7.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
discord==1.0.1
discord.py>=1.3.4
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
oauthlib==3.1.0
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
requests>=2.24.0
requests-oauthlib>=1.3.0
rsa==4.7
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3>=1.26.5
websockets==9.1
wrapt==1.12.1
yarl==1.5.0
